refactor(socket): mock 기능은 mock 모드에서만 노출
