+++ socket-service/src/main/kotlin/com/parkit/socket/mock/CoachingScheduler.kt
@@ -1,4 +1,4 @@
-package com.parkit.socket.controller
+package com.parkit.socket.mock
 
 import com.parkit.socket.dto.CoachingSocketDto
 import com.parkit.socket.dto.ObstacleDistancesDto
@@ -6,25 +6,23 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.messaging.simp.SimpMessagingTemplate
 import org.springframework.scheduling.annotation.EnableScheduling
 import org.springframework.scheduling.annotation.Scheduled
-import org.springframework.stereotype.Controller
+import org.springframework.stereotype.Component
 import kotlin.random.Random
 
-@Controller
+@Component
 @EnableScheduling
 @ConditionalOnProperty(prefix = "parkit.mock.coaching", name = ["enabled"], havingValue = "true")
-class MockCoachingController(
-    private val messagingTemplate: SimpMessagingTemplate
+class CoachingScheduler(
+	private val messagingTemplate: SimpMessagingTemplate,
 ) {
-
 	companion object {
 		private const val TOPIC_DESTINATION = "/topic/coaching-mock"
 		private const val DANGER_LIMIT_FRONT_BACK = 200
 		private const val DANGER_LIMIT_SIDE = 80
 	}
 
-    // 1초마다 가짜 주차 코칭 데이터를 브로드캐스트
-    @Scheduled(fixedRate = 1000)
-    fun sendMockCoachingData() {
+	@Scheduled(fixedRate = 1000)
+	fun sendMockCoachingData() {
 		val frontGap = Random.nextInt(50, 301)
 		val backGap = Random.nextInt(50, 301)
 		val leftGap = Random.nextInt(30, 151)
@@ -54,7 +52,6 @@ class MockCoachingController(
 			coachingId = coachingId,
 		)
 
-        // 브로커의 /topic/coaching 을 구독하는 클라이언트들에게 전달
 		messagingTemplate.convertAndSend(TOPIC_DESTINATION, mockData)
-    }
+	}
 }
